Properly handle name lengths of (unsigned int)-1
When a user sends a name length value of 0xffffffff, nbd-server will try
to allocate one more byte for the \0 at the end, but that will result in
an integer overflow and a malloc(0), with the resulting write being to a
dangling pointer.

Fix by constraining the string size to 4096 bytes, as recommended by the
protocol standard.

This issue exists in NBD_OPT_INFO/NBD_OPT_GO handling as well as in
NBD_OPT_EXPORT_NAME handling.

CVE-2022-26495

Reported-By: 王多 <duo.wang@chaitin.com>
Signed-Off-By: Wouter Verhelst <w@uter.be>