Optionally use cwrap
We currently conflict with a system-wide NBD implementation (if one is
running).

Optionally LD_PRELOAD the cwrap libraries in the test suite, if found,
so that we don't do that anymore.

In future we might perhaps also add extra tests that do more if the cwrap
libraries are available, but not now.