docs: Clarify server behavior on read failures
Back in 3.14, we clarified the documentation to state that a read
error detected before the header is sent must use a 0-length payload
to keep the connection alive, to match most existing server and client
implementations, rather than initiating a hard disconnect or sending a
full length payload of garbage.  We then tweaked another spot in the
docs to match in 3.17 (commit e5474502), but then regressed when
merging in support for structured replies shortly after, since that
was developed in parallel.

Fixes: d1fcc84b ("Merge branch 'origin' into extension-structured-reply", 3.17)