Refactor request handling
Currently, the state of a request is stored on the stack, with the reply
being built in a buffer which is sent out at the very end of the
'handle_normal_read' function.

This makes implementing structured replies complicated, as for those we
may want to sometimes send a reply chunk before that point.

Thus, refactor request handling such that we no longer depend on the
stack for sending out things.

Signed-off-by: Wouter Verhelst <w@uter.be>