docs: Prefer 'cookie' over 'handle'
In libnbd, we quickly learned that distinguishing between 'handle'
(verb for acting on an object) and 'handle' (noun describing which
object to act on) could get confusing; we solved it by renaming the
latter to 'cookie'.  Copy that approach into the NBD spec, and make it
obvious that a cookie is opaque data from the point of view of the
server.  Makes no difference to implementations (other than older code
still using 'handle' may be slightly harder to tie back to the spec).

In parallel, a patch has been proposed to update the kernel uapi to
prefer the same naming [1], as well as preferring operations on a
64-bit integer rather than a char[8] (doesn't matter in practice when
the value is opaque, but does make it easier to write wireshark
sniffers that can expect a certain endianness for correlation back to
log messages when the logs also use 64-bit integers).

[1] https://lkml.org/lkml/2023/3/17/1107

Suggested-by: Richard W.M. Jones <rjones@redhat.com>
Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20230309210623.2238671-2-eblake@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>