nbd: Use uint64_t instead of char[8] for cookie
Type-punning *(uint64_t*)(char[8]) is unsafe if the pointer is not
aligned per the requirements of the hardware (x86_64 has 1-byte
alignment, but other hardware exists that will give a SIGBUS).  In
practice, it didn't matter - 'struct nbd_request' was already 64-bit
aligned for 'from' (even before the recent change in commit 739f7121
to pack it), and 'struct nbd_reply' being packed means the compiler
emits code to deal with 1-byte alignment despite hardware.  But since
the cookie is already opaque on the server side, we might as well
treat it as an 8-byte integer instead of a character array, with no
visible semantic changes to the client.

This patch does NOT perform endian swapping on the value.  As long as
we are consistent (no swap when reading from the client, no swap when
writing back to the client), the client sees the same value.  If we
were to log the cookie at any point in our local handling, then it
would be nice to obey the spec that all NBD values sent over the wire
are in network byte order (big-endian), so that what we log appears as
a 64-bit value with endianness matching whatever wireshark or the
client might be logging as well.  But since we aren't logging the
cookie, skipping the byte-swaps doesn't hurt.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20230309210623.2238671-4-eblake@redhat.com>