uapi nbd: improve doc links to userspace spec
On Fri, Mar 10, 2023 at 02:15:23PM -0600, Eric Blake wrote:
> The documentation file also had a stale link to sourceforge; nbd
> ditched that several years ago in favor of github.

While it's accurate to say that we've mostly moved to github, I wouldn't
describe the sourceforge link as "stale"; I've kept the information on
that landing page generally up-to-date over the years (the most recent
change was the addition of the link to Ceph for Windows for a Windows
NBD client, in 2020), and I do still post file releases on sourceforge,
primarily.

However, I guess it's been long enough, and there isn't really anything
that sourceforge can provide us which github can't, so I suppose we
might as well.