Drop support for Sockets Direct Protocol (SDP)
The SDP support in nbd-client was a drive-by patch originally given to us way
back in 2007 (before git!), and not revisited since.

According to Wikipedia,

  This protocol has been deprecated[5] or not maintained by OFED in the latest
  releases

and also, according to
https://lists.openfabrics.org/pipermail/ewg/2012-April/017365.html,

  > So where did it go? Is SDP obsolete?

  its deprecated

So, as I don't have the hardware required to test these patches, and as
I also don't remember anyone ever stepping forward with any comments or
remarks about the SDP support in the 15 years since we added this, and
since apparently this is now deprecated, let's just drop it.

I considered keeping support for the command-line options and/or the
config file options, but I think that isn't all that useful.

Signed-off-by: Wouter Verhelst <w@uter.be>