spec: Introduce NBD_FLAG_BLOCK_STATUS_PAYLOAD
NBD_CMD_BLOCK_STATUS currently forces the server to reply to all
metacontext ids that the client negotiated via
NBD_OPT_SET_META_CONTEXT.  But since extended headers make it easy for
the client to pass command payloads, we can allow for a client to
negotiate multiple metacontexts up front but express dynamic interest
in varying subsets of those contexts over the life of the connection,
for less wasted effort in responding to NBD_CMD_BLOCK_STATUS.  This
works by having the command payload supply an effect length and a list
of ids the client is currently interested in.

Note that this is an optional extension, not mandated by extended
headers proper.  That is because a server may only support a single
metacontext (typically "base:allocation"), at which point the
implementation overhead of supporting a client payload to filter out
which context the client is interested in just adds work.  Hence, it
also requires the server to advertise a bit to the client to know when
the server does support request filtering, rather than the client
relying on the extended headers negotiation.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20230413220241.1396012-6-eblake@redhat.com>
Reviewed-By: Wouter Verhelst <w@uter.be>