spec: Add NBD_OPT_EXTENDED_HEADERS
Add a new negotiation feature where the client and server agree to use
larger packet headers on every packet sent during transmission phase.
This has two purposes: first, it makes it possible to perform
operations like trim, write zeroes, and block status on more than 2^32
bytes in a single command.  While at it, this is a great opportunity
to declare that any client opting in to extended headers will use
NBD_OPT_GO instead of the weaker NBD_OPT_EXPORT_NAME, in part because
that encourages both server and client to communicate size constraints
during handshaking (although advertising any additional constraints
would be a separate extension, to be added later).

Second, when structured replies are active, clients have to deal with
the difference between 16- and 20-byte headers of simple
vs. structured replies, which impacts performance if the client must
perform multiple syscalls to first read the magic before knowing if
there are even additional bytes to read to learn a payload length.  In
extended header mode, all headers are the same width and there are no
simple replies permitted.  The layout of the reply header is more like
the request header; and echoing the client's offset in the reply makes
it easier to convert between absolute and relative offsets for replies
to NBD_CMD_READ.  Similarly, by having extended mode use a power-of-2
sizing, it becomes easier to manipulate arrays of headers without
worrying about an individual header crossing a cache line.  However,
note that this change only affects the headers; data payloads can
still be unaligned (for example, a client performing 1-byte reads or
writes).  We would need to negotiate yet another extension if we
wanted to ensure that all NBD transmission packets started on an
8-byte boundary after option haggling has completed.

Note that NBD_INFO_BLOCK_SIZE still advertises only a 32-bit value for
maximum payload size.  As of this patch, it is assumed that most
clients will still be limiting NBD_CMD_READ requests to a much smaller
cap of 32M (the portable payload limit if nothing was advertised) or
64M (nbdkit chose this limit, and a client may know it is talking to
an nbdkit server even without explicit advertisement); so keeping a
32-bit payload size is still fine even while benefittng from a 64-bit
effect size.  At any rate, this patch intentionally does not worry
about a client and server that can mutually agree to a maximum payload
larger than 4G by some means other than NBD_INFO_BLOCK_SIZE (we can
standardize it as a future extension if someone demonstrates a use
case for it); meanwhile, NBD_REPLY_TYPE_OFFSET_DATA will work as-is
with larger payload (since its length is part of the extended header),
but for holes, a server would have to either add a 64-bit counterpart
to NBD_REPLY_TYPE_OFFSET_HOLE or split its hole responses over
multiple reply chunks.

But for NBD_CMD_BLOCK_STATUS, it is very useful to be able to report a
status extent with a size larger than 32-bits, in some cases even if
the client's request was for an effect length smaller than 32-bits
(such as when it is known that the entire image is not sparse).
Although it would techincally be possible to let a server choose
between 32- and 64-bit status replies on a per-response basis, the
consensus was that as long as extended headers are negotiated, it is
fewer corner cases to worry about by mandating in this patch that the
server always answers with the new `NBD_REPLY_TYPE_BLOCK_STATUS_EXT`.

This spec addition was done in parallel with proof of concept
implementations in qemu (server and client) and libnbd (client); but
for now will be left on an extension branch util there is also an
implementation in nbdkit (server).

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20230413220241.1396012-4-eblake@redhat.com>