nbd-server: do not straddle uses of GThreadPool across a fork()
We created the thread pool at the main initialization, before forking
off a child. This used to work just fine, but as of GLib 2.76, this no
longer works due to changes internal to the implementation of
GThreadPool.

Since we don't need to use the thread pool before the fork() call
anyway, stop trying to do so and avoid the problem altogether.

Closes: gh-146
Signed-off-by: Wouter Verhelst <w@uter.be>