server: Avoid sync bug when NBD_OPT_GO fails with ERR_POLICY
It is imperative that all client data is consumed when replying with
error to NBD_OPT_GO, so that the server will still be in sync if the
client attempts another NBD_OPT_* command.  We got this right a couple
lines earlier when server is NULL, but leave unread client data when
commit_client() fails.

Fixes: e6b56c12 ("Merge branch 'extension-info'", 3.16)
Signed-off-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Wouter Verhelst <w@uter.be>