Fix the check & no cases of enable_manpages
Currently, running "configure --disable-manpages" while docbook2man *is*
installed results in the error "don't know what to do here" when it
should disable manpages.

There also appears to be a missing conditional at the start of the line;
there's closing un-matched ]) at the end of the line. Still, at this
point the check can be done in pure shell; no need for AC macros. I've
also removed the confusing m4_divert_text call on the check case. Not
sure why that was there, but it appears unnecessary.