Add docbook build-dep, re-arrange build-deps
The docbook build-dep is technically required as it contains the DTDs
for docbook2man to build manpages. It is implicitly pulled in under
sbuild currently because docbook-utils deps on docbook-dsssl which deps
on (docbook | docbook-xml). However, if docbook-xml is installed, the
latter dep is satisified, docbook isn't installed and the build fails.

I also took the liberty of breaking the build-deps into separate lines
for readability -- happy to revert that change if it's not desired!