Fix lintian warning: missing-systemd-service-for-init.d-script
Add a native systemd unit definition for the nbd-server service. This
defines it as a classic "forking" type service, explicitly specifies the
PID file location, and sets it up to start after the network-online
target.

This should probably be tightened up with some of the security /
isolation facilities that systemd provides but this is just a basic
translation of the existing auto-generated unit.