Apparently these tests can't co-exist
autopkgtest hilariously *doesn't* reset the testbed between tests if
their deps are equivalent (or more precisely if next-test's deps are a
superset of current-test's deps). As a result, because these tests have
equivalent deps, they wind up treading on each other's configurations.

Given the read-write test is more or less a superset of
nbdtab-ipv4-address anyway we may as well make it a proper superset by
tweaking the nbdtab to use an IPv4 style address.