Refactor the negotiate() and connected functions
Having a gazillion arguments to negotiate is unwieldy and
unmaintainable. Use the CLIENT type to handle parsed stuff instead.

Originally we were going to pass a pointer to a CLIENT, but we need to
keep the cur_client global variable for config file parsing, so might
as well reuse that.

Not so clean, so we might revisit this in the future, but at least this
reduces complexity somewhat.

Signed-off-by: Wouter Verhelst <w@uter.be>