nbd-client: Add support for setting /sys/block/nbdN/backend
In order for userspace apps to idempotently ensure that a
netlink-managed NBD device corresponds to the device that the app is
expecting, kernel 5.14 added a netlink backend string identifier.
Expose the ability for setting this identifier when binding to a
device; the user can then check with /sys/block/nbdN/backend to see if
the contents match expectations.

Signed-off-by: Eric Blake <eblake@redhat.com>