spec: Relax block status alignment to match existing servers
At least nbdkit 1.42 has several scenarios where it can advertise a
minimum block size, but where block status results are not aligned to
that size.  While most of those instances are bugs fixed in the
upcoming 1.44, we have to consider the case when a server advertises
an image size which is not a multiple of the minimum block size.  The
spec is already clear that a server SHOULD advertise aligned sizes,
but when it doesn't, the requirement that block status results be
aligned is impossible to meet.  Relaxing the standard from MUST to
SHOULD warns clients to be prepared for weaknesses in the server, as
well as making it less troublesome to try and collect block status
even for an unaligned tail of an image.

Signed-off-by: Eric Blake <eblake@redhat.com>